Fixed warnings for -std=c++11. Added CStrDup function to replace strdup because it's a POSIX function that isn't part of the standard. Made ScrollBar::SetScrollWidgets allow you to reset the default widgets. Other minor changes. More warnings enabled now.

+++ include/Eagle/Gui/Scroller.hpp
@@ -49,7 +49,7 @@ protected :
    bool mouse_drag_handle;
    int start_mouse_x;
    int start_mouse_y;
-   float start_percent;
+   int start_percent;
    
    bool down;
    int repeat_scroll_increment;
